% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AirSensor2-package.R
\docType{package}
\name{AirSensor2-package}
\alias{AirSensor2-package}
\title{AirSensor2: Process and Display Data from Air Quality Sensors}
\description{
Process and display data from low-cost air quality sensors. Initial focus is on PM2.5 measurements from sensors produced by 'PurpleAir' \url{https://www2.purpleair.com}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/MazamaScience/AirSensor2}
  \item Report bugs at \url{https://github.com/MazamaScience/AirSensor2/issues}
}

}
\author{
\strong{Maintainer}: Jonathan Callahan \email{jonathan.s.callahan@gmail.com}

}
\keyword{internal}
