% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Clarity_API.R
\name{Clarity_getAllOpenHourly}
\alias{Clarity_getAllOpenHourly}
\title{Retrieve current hourly data for all Open Data sensors.}
\usage{
Clarity_getAllOpenHourly(
  api_key = NULL,
  format = c("USFS2", "USFS"),
  baseUrl =
    "https://clarity-data-api.clarity.io/v1/open/all-recent-measurement/pm25/hourly"
)
}
\arguments{
\item{api_key}{Clarity API READ Key. If `api_key = NULL`, it
will be obtained using `getAPIKey("Clarity-read")`.}

\item{format}{Customized output format ("USFS2", "USFS").}

\item{baseUrl}{URL endpoint.}
}
\value{
List containing five data frames: `synoptic`, `pm2.5_QCFlag`,
`pm2.5`, `nowcast_QCFlag` and `nowcast`.
}
\description{
Sends a request to the Clarity API endpoint for Open Data.

When `format = "USFS2"`, two additional fields are returned in the
"synoptic" dataframe: `calibrationId` and `calibrationCategory`.

\itemize{
  \item{*Measurement data from* -- All open datasources}
  \item{*Measurements returned* -- Hourly values for last 3 hours}
  \item{*PM2.5 Mass Concentration* -- 1-Hour Mean Nowcast}
}
}
\examples{
\donttest{
# Fail gracefully if any resources are not available
try({

library(AirSensor2)

  Clarity_getAllOpenHourly(
    api_key = Clarity_API_READ_KEY,
    format = "USFS2"
  )

}, silent = FALSE)
}
}
