% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Clarity_API.R
\name{Clarity_getOpenHourly}
\alias{Clarity_getOpenHourly}
\title{Retrieve current hourly data from a single source.}
\usage{
Clarity_getOpenHourly(
  api_key = NULL,
  datasourceId = NULL,
  format = "USFS2",
  baseUrl = "https://clarity-data-api.clarity.io/v1/open/datasource-measurement"
)
}
\arguments{
\item{api_key}{Clarity API READ Key. If `api_key = NULL`, it
will be obtained using `getAPIKey("Clarity-read")`.}

\item{datasourceId}{Clarity sensor identifier.}

\item{format}{Customized output format ("USFS2", "USFS").}

\item{baseUrl}{URL endpoint.}
}
\value{
List containing five data frames: `meta`, `pm2.5_QCFlag`,
`pm2.5`, `nowcast_QCFlag` and `nowcast`.
}
\description{
Sends a request to the Clarity API endpoint for Open Data.

When `format = "USFS2"`, two additional fields are returned in the
"meta" dataframe: `calibrationId` and `calibrationCategory`.

\itemize{
  \item{*Measurement data from* -- Drill down on one open datasource}
  \item{*Measurements returned* -- Hourly values  for last 10 days}
  \item{*PM2.5 Mass Concentration* -- 1-Hour Mean Nowcast}
}
}
\examples{
\donttest{
# Fail gracefully if any resources are not available
try({

library(AirSensor2)

  Clarity_getOpenHourly(
    api_key = Clarity_API_READ_KEY,
    datasourceId = "DAABL1560",
    format = "USFS2"
  )

}, silent = FALSE)
}
}
