% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\encoding{UTF-8}
\name{example_pas}
\alias{example_pas}
\title{Example enhanced Purple Air Synoptic dataset}
\format{
A tibble with 1948 rows and 50 columns of data.
}
\source{
https://www2.purpleair.com
}
\usage{
example_pas
}
\description{
The \code{example_pas} dataset provides a quickly loadable
version of a \emph{pas} object for practicing and code examples.
This dataset contains data for sensors in Washington and Oregon
and was generated on 2023-02-15 by running:

\preformatted{
library(AirSensor2)

initializeMazamaSpatialUtils()

example_pas <-
  pas_createNew(
    api_key = PURPLE_AIR_API_READ_KEY,
    countryCodes = "US",
    stateCodes = c("WA", "OR"),
    counties = NULL,
    lookbackDays = 1,
    location_type = 0
  )

save(example_pas, file = "data/example_pas.rda")
}
}
\seealso{
example_pas_raw
}
\keyword{datasets}
