% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\encoding{UTF-8}
\name{example_pas_historical}
\alias{example_pas_historical}
\title{Example enhanced Purple Air Synoptic dataset}
\format{
A tibble with 2358 rows and 51 columns of data.
}
\source{
https://www2.purpleair.com
}
\usage{
example_pas_historical
}
\description{
The \code{example_pas} dataset provides a quickly loadable
version of a \emph{pas} object for practicing and code examples.
This dataset contains data for sensors in Washington and Oregon
and was generated on 2023-11-29 by running:

\preformatted{
library(AirSensor2)

initializeMazamaSpatialUtils()

source("global_vars.R") # contains PurpleAir_API_READ_KEY

example_pas_historical <-
  pas_createNew(
    api_key = PurpleAir_API_READ_KEY,
    fields = PurpleAir_SENSOR_METADATA_FIELDS,
    countryCodes = "US",
    stateCodes = "WA",
    counties = "Okanogan",
    lookbackDays = 365 * 10,     # 10 years
    location_type = 0            # Outdoor only
  )

save(example_pas_historical, file = "data/example_pas_historical.rda")
}
}
\seealso{
example_pas_raw
}
\keyword{datasets}
