% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\encoding{UTF-8}
\name{example_pas_metadata}
\alias{example_pas_metadata}
\title{Example enhanced PurpleAir Synoptic dataset}
\format{
A tibble with 122 rows and 31 columns of data.
}
\source{
https://www2.purpleair.com
}
\usage{
example_pas_metadata
}
\description{
The \code{example_pas_metadata} dataset provides a quickly loadable
version of a \emph{pas} object for practicing and code examples. This
object contains more metadata than other \code{example_pas_~} objects.

This dataset contains data for sensors in Washington and Oregon
and was generated on 2024-03-12 by running:

\preformatted{
library(AirSensor2)

initializeMazamaSpatialUtils()

source("global_vars.R") # contains PurpleAir_API_READ_KEY

example_pas_metadata <-
  pas_createNew(
    api_key = PurpleAir_API_READ_KEY,
    fields = PurpleAir_PAS_METADATA_FIELDS,
    countryCodes = "US",
    stateCodes = "WA",
    counties = "Okanogan",
    lookbackDays = 0,            # all years
    location_type = 0            # outdoor only
  )

save(example_pas_metadata, file = "data/example_pas_metadata.rda")
}
}
\seealso{
example_pas_pm25

example_pas_historical
}
\keyword{datasets}
