% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\encoding{UTF-8}
\name{example_pas_pm25}
\alias{example_pas_pm25}
\title{Example enhanced PurpleAir Synoptic dataset}
\format{
A tibble with 2287 rows and 33 columns of data.
}
\source{
https://www2.purpleair.com
}
\usage{
example_pas_pm25
}
\description{
The \code{example_pas} dataset provides a quickly loadable
version of a \emph{pas} object for practicing and code examples.
This dataset contains data for sensors in Washington and Oregon
and was generated on 2024-03-12 by running:

\preformatted{
library(AirSensor2)

initializeMazamaSpatialUtils()

source("global_vars.R") # contains PurpleAir_API_READ_KEY

example_pas_pm25 <-
  pas_createNew(
    api_key = PurpleAir_API_READ_KEY,
    fields = PurpleAir_PAS_AVG_PM25_FIELDS,
    countryCodes = "US",
    stateCodes = c("WA", "OR"),
    counties = NULL,
    lookbackDays = 1,
    location_type = 0
  )

save(example_pas_pm25, file = "data/example_pas_pm25.rda")
}
}
\seealso{
example_pas_pm25_raw

example_pas_historical

example_pas_metadata
}
\keyword{datasets}
