% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initializeMazamaSpatialUtils.R
\name{initializeMazamaSpatialUtils}
\alias{initializeMazamaSpatialUtils}
\title{Initialize MazamaSpatialUtils}
\usage{
initializeMazamaSpatialUtils(
  spatialDataDir = "~/Data/Spatial",
  stateCodeDataset = "NaturalEarthAdm1.rda",
  USCountiesDataset = "USCensusCounties.rda",
  logLevel = WARN
)
}
\arguments{
\item{spatialDataDir}{Directory containing external
\pkg{MazamaSpatialUtils} spatial datasets.}

\item{stateCodeDataset}{Name of the \pkg{MazamaSpatialUtils} dataset
containing state or administrative-level boundaries and codes.}

\item{USCountiesDataset}{Name of the \pkg{MazamaSpatialUtils} dataset
containing US county names.}

\item{logLevel}{Logging level to use if logging has not already been
initialized.}
}
\value{
Invisible `NULL`.
}
\description{
Convenience function that loads the spatial datasets needed by
any of the "enhanceRaw" functions.

This function wraps the equivalent sequence:

\preformatted{
  data("SimpleCountriesEEZ", package = "MazamaSpatialUtils")
  data("SimpleTimezones", package = "MazamaSpatialUtils")

  MazamaSpatialUtils::setSpatialDataDir("~/Data/Spatial")
  MazamaSpatialUtils::loadSpatialData("NaturalEarthAdm1.rda")
  MazamaSpatialUtils::loadSpatialData("USCensusCounties.rda")
}

It should be run before using `pas_load()`, because `pas_load()` uses the
spatial data loaded here to enhance raw synoptic data via
`pas_enhanceData()`.

If file logging is desired, the individual logging commands should be run
separately with output files specified in `MazamaCoreUtils::logger.setup()`.
}
