% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OpenAQ_createSynoptic.R
\name{OpenAQ_createLocations}
\alias{OpenAQ_createLocations}
\title{Create a new OpenAQ locations dataset}
\usage{
OpenAQ_createLocations(
  api_key = NULL,
  countryCodes = NULL,
  stateCodes = NULL,
  counties = NULL,
  lookbackDays = 1,
  providers = NULL,
  manufacturers = NULL,
  monitor = NULL,
  limit = 1000
)
}
\arguments{
\item{api_key}{OpenAQ API READ Key. If `api_key = NULL`, it
will be obtained using `getAPIKey("OPENAQ")`.}

\item{countryCodes}{ISO 3166-1 alpha-2 country codes used to subset the data.
At least one countryCode must be specified. (Country names are also acceptible.)}

\item{stateCodes}{ISO-3166-2 alpha-2 state codes used to subset the data.
Specifying stateCodes is optional.}

\item{counties}{US county names or 5-digit FIPS codes used to subset the data.
Specifying counties is optional.}

\item{lookbackDays}{Number of days to "look back" for locations with valid data. Locations are
filtered to only include those with data more recent than `lookbackDays` ago.
Use `lookbackDays = 0` to get all historical locations.}

\item{providers}{Character vector of provider names or export prefixes used to subset the data. See note.}

\item{manufacturers}{Character vector of manufacturer names used to subset the data. See note.}

\item{monitor}{A logical to filter results to regulatory monitors (TRUE) or air sensors (FALSE), both are included if NULL.}

\item{limit}{An integer specifying the maximum number of results to return, max is 1000.}
}
\value{
A tibble of location metadata.
}
\description{
Download, parse and enhance locations data from OpenAQ and
return the results as a plain tibble.

Steps include:

1) Download and parse locations data

2) Add spatial metadata for each location including:
\itemize{
  \item{timezone -- olson timezone}
  \item{countryCode -- ISO 3166-1 alpha-2}
  \item{stateCode -- ISO 3166-2 alpha-2}
}
}
\note{
Valid values for `countryCodes`, `providers` and `manufacturers` can be obtained with:

\preformatted{
OpenAQ_getCountries()
OpenAQ_getProviders()
OpenAQ_getManufacturers()
}

These functions return reference tables containing the available values.
Matching is case-insensitive. Unmatched values are ignored.
}
\examples{
\donttest{
# Fail gracefully if any resources are not available
try({

library(AirSensor2)

initializeMazamaSpatialUtils()

synoptic <-
  OpenAQ_createLocations(
    countryCodes = "US",
    stateCodes = "IL",
    counties = "Cook",
    manufacturers = "Clarity",
    api_key = OPENAQ_API_KEY
  )

synoptic \%>\% MazamaLocationUtils::table_leaflet()

}, silent = FALSE)
}
}
