% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OpenAQ_createMonitor.R
\name{OpenAQ_createMonitor}
\alias{OpenAQ_createMonitor}
\title{Create a new OpenAQ 'mts_monitor' object}
\usage{
OpenAQ_createMonitor(
  locations = NULL,
  locations_id = NULL,
  parameter = c("pm25"),
  startdate = NULL,
  enddate = NULL,
  timezone = "UTC",
  applyQC = TRUE,
  maxPages = 1,
  sleepSeconds = 0.2,
  api_key = NULL
)
}
\arguments{
\item{locations}{Previously generated locations object containing `id`.}

\item{locations_id}{OpenAQ location identifier.}

\item{parameter}{Parameter to use for data. Currently only `"pm25"` is supported.}

\item{startdate}{Start datetime in the requested timezone.}

\item{enddate}{End datetime in the requested timezone.}

\item{timezone}{Olson timezone used to interpret `startdate` and `enddate`.}

\item{applyQC}{Logical specifying whether to apply basic QC to invalidate
bad data values.}

\item{maxPages}{Maximum number of pages to request automatically.}

\item{sleepSeconds}{Number of seconds to pause between additional requests.}

\item{api_key}{OpenAQ API key. If `api_key = NULL`, it
will be obtained using `getAPIKey("OpenAQ-read")`.}
}
\value{
An AirMonitor package `mts_monitor` object.
}
\description{
Download, parse and enhance hourly timeseries data from OpenAQ
and create an object of class `mts_monitor` for use with the AirMonitor
package.
}
