% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OpenAQ_downloadRawLocations.R
\name{OpenAQ_downloadRawLocations}
\alias{OpenAQ_downloadRawLocations}
\title{Download raw OpenAQ location data}
\usage{
OpenAQ_downloadRawLocations(
  api_key = NULL,
  bbox = NULL,
  providers_id = NULL,
  manufacturers_id = NULL,
  monitor = NULL,
  mobile = FALSE,
  countries_id = NULL,
  limit = 1000,
  maxPages = 1,
  sleepSeconds = 0.2
)
}
\arguments{
\item{api_key}{OpenAQ API key.}

\item{bbox}{Bounding box passed to [openaq::list_locations()].}

\item{providers_id}{Provider IDs passed to [openaq::list_locations()].}

\item{manufacturers_id}{Manufacturer IDs passed to [openaq::list_locations()].}

\item{monitor}{Logical value passed to [openaq::list_locations()].}

\item{mobile}{Logical value passed to [openaq::list_locations()].}

\item{countries_id}{Country IDs passed to [openaq::list_locations()].}

\item{limit}{Maximum number of records to request per page.}

\item{maxPages}{Maximum number of pages to request automatically.}

\item{sleepSeconds}{Number of seconds to pause between additional requests.}
}
\value{
A dataframe of raw OpenAQ location metadata.
}
\description{
Downloads raw location metadata from OpenAQ and optionally retrieves
additional pages when many matching locations are found.
}
\details{
When more matching locations are available than can be returned in a single
request, this function can automatically request additional pages up to
`maxPages`. If more pages may still be available after `maxPages` pages have
been retrieved, a warning is issued suggesting that the user narrow the
request.
}
