% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OpenAQ_API.R
\name{OpenAQ_getCountries}
\alias{OpenAQ_getCountries}
\title{Get OpenAQ country metadata}
\usage{
OpenAQ_getCountries()
}
\value{
A tibble of OpenAQ country metadata.
}
\description{
Returns OpenAQ country metadata as a tibble. Data are downloaded only once
per R session and then cached for re-use.
}
\details{
This function requires the optional package `openaq`.
}
