% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OpenAQ_API.R
\name{OpenAQ_getInstruments}
\alias{OpenAQ_getInstruments}
\title{Get OpenAQ instrument metadata}
\usage{
OpenAQ_getInstruments()
}
\value{
A tibble of OpenAQ instrument metadata.
}
\description{
Returns OpenAQ instrument metadata as a tibble. Data are downloaded only once
per R session and then cached for re-use.
}
\details{
This function requires the optional package `openaq`.
}
