% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OpenAQ_API.R
\name{OpenAQ_getLicenses}
\alias{OpenAQ_getLicenses}
\title{Get OpenAQ license metadata}
\usage{
OpenAQ_getLicenses()
}
\value{
A tibble of OpenAQ license metadata.
}
\description{
Returns OpenAQ license metadata as a tibble. Data are downloaded only once
per R session and then cached for re-use.
}
\details{
This function requires the optional package `openaq`.

For more information about the OpenAQ data access API, see:
\url{https://docs.openaq.org/about/about}
}
