% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OpenAQ_API.R
\name{OpenAQ_getManufacturers}
\alias{OpenAQ_getManufacturers}
\title{Get OpenAQ manufacturer metadata}
\usage{
OpenAQ_getManufacturers()
}
\value{
A tibble of OpenAQ manufacturer metadata.
}
\description{
Returns OpenAQ manufacturer metadata as a tibble. Data are downloaded only
once per R session and then cached for re-use.
}
\details{
This function requires the optional package `openaq`.

For more information about the OpenAQ data access API, see:
\url{https://docs.openaq.org/about/about}
}
