% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OpenAQ_API.R
\name{OpenAQ_getParameters}
\alias{OpenAQ_getParameters}
\title{Get OpenAQ parameter metadata}
\usage{
OpenAQ_getParameters()
}
\value{
A tibble of OpenAQ parameter metadata.
}
\description{
Returns OpenAQ parameter metadata as a tibble. Data are downloaded only once
per R session and then cached for re-use.
}
\details{
This function requires the optional package `openaq`.
}
