% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OpenAQ_API.R
\name{OpenAQ_getProviders}
\alias{OpenAQ_getProviders}
\title{Get OpenAQ provider metadata}
\usage{
OpenAQ_getProviders()
}
\value{
A tibble of OpenAQ provider metadata.
}
\description{
Returns OpenAQ provider metadata as a tibble. Data are downloaded only once
per R session and then cached for re-use.
}
\details{
This function requires the optional package `openaq`.

For more information about the OpenAQ data access API, see:
\url{https://docs.openaq.org/about/about}
}
