% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OpenAQ_API.R
\name{OpenAQ_instrumentToID}
\alias{OpenAQ_instrumentToID}
\title{Convert instrument names to OpenAQ instrument IDs}
\usage{
OpenAQ_instrumentToID(x)
}
\arguments{
\item{x}{Character vector of instrument names.}
}
\value{
Integer vector of OpenAQ instrument IDs.
}
\description{
Accepts a character vector of instrument names and returns the corresponding
OpenAQ instrument IDs.
}
\details{
Matching is case-insensitive. If no match is found, `NA` is returned.
}
\examples{
\dontrun{
OpenAQ_instrumentToID(c("BAM 1020", "PurpleAir PA-II", "Unknown"))
}

}
