% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OpenAQ_API.R
\name{OpenAQ_licenseToID}
\alias{OpenAQ_licenseToID}
\title{Convert license names to OpenAQ license IDs}
\usage{
OpenAQ_licenseToID(x)
}
\arguments{
\item{x}{Character vector of license names.}
}
\value{
Integer vector of OpenAQ license IDs.
}
\description{
Accepts a character vector of license names and returns the corresponding
OpenAQ license IDs.
}
\details{
Matching is case-insensitive. If no match is found, `NA` is returned.
}
\examples{
\dontrun{
OpenAQ_licenseToID(c("CC BY 4.0", "Public Domain", "Unknown"))
}

}
