% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OpenAQ_API.R
\name{OpenAQ_manufacturerToID}
\alias{OpenAQ_manufacturerToID}
\title{Convert manufacturer names to OpenAQ manufacturer IDs}
\usage{
OpenAQ_manufacturerToID(x)
}
\arguments{
\item{x}{Character vector of manufacturer names.}
}
\value{
Integer vector of OpenAQ manufacturer IDs.
}
\description{
Accepts a character vector of manufacturer names and returns the corresponding
OpenAQ manufacturer IDs.
}
\details{
Matching is case-insensitive. If no match is found, `NA` is returned.
}
\examples{
\dontrun{
OpenAQ_manufacturerToID(c("Met One Instruments", "Thermo Fisher Scientific", "Unknown"))
}

}
