% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pas_downloadParseRawData_2020.R
\name{pas_downloadParseRawData_2020}
\alias{pas_downloadParseRawData_2020}
\title{Download synoptic data from PurpleAir with older API}
\usage{
pas_downloadParseRawData_2020(
  baseUrl = "https://www.purpleair.com/json?all=true"
)
}
\arguments{
\item{baseUrl}{Base URL for access to PurpleAir synoptic data. This is the
old, 2020 API for accessing synoptic data tha may disappear at any point.}
}
\value{
Dataframe of synoptic PurpleAir data.
}
\description{
Download and parse synoptic data for the PurpleAir network
of particulate sensors.

The synoptic data provides a view of the entire PurpleAir network and
includes both metadata and recent PM2.5 averages for each deployed sensor.
}
\examples{
\donttest{
# Fail gracefully if any resources are not available
try({

library(AirSensor2)

pas_raw <- pas_downloadParseRawData_2020()

if ( interactive() ) View(pas_raw[1:100,])

}, silent = FALSE)
}
}
\references{
\href{https://www.purpleair.com/json?all=true}{json formatted PurpleAir data}
}
