% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pas_enhanceRawData.R
\name{pas_enhanceRawData}
\alias{pas_enhanceRawData}
\title{Enhance synoptic data from PurpleAir}
\usage{
pas_enhanceRawData(pas_raw = NULL, countryCodes = NULL, stateCodes = NULL)
}
\arguments{
\item{pas_raw}{Dataframe returned by \code{pas_downloadParseRawData()}.}

\item{countryCodes}{ISO 3166-1 alpha-2 country codes used to subset the data.}

\item{stateCodes}{ISO-3166-2 alpha-2 state codes used to subset the data.}
}
\value{
Enhanced dataframe of synoptic PurpleAir data.
}
\description{
Enhance raw synoptic data from PurpleAir to create an improved
dataframe compatible with the \pkg{MazamaLocationUtils} package.

Steps include:

1) Replace variable with more consistent, human readable names.

2) Add spatial metadata for each sensor including:
\itemize{
  \item{timezone -- Olson timezone}
  \item{countryCode -- ISO 3166-1 alpha-2}
  \item{stateCode -- ISO 3166-2 alpha-2}
}

3) Convert data types from character to \code{POSIXct} and \code{numeric}.

4) Add additional metadata items:
\itemize{
\item{sensorManufacturer = "Purple Air"}
}

Filtering by country can speed up the process of enhancement and may be
performed by providing a vector ISO country codes to the \code{countryCodes}
argument. By default, no subsetting is performed.

Uses may also limit results by specifying \code{stateCodes} when
\code{countryCodes} is limited to a single country.
}
\references{
\href{https://api.purpleair.com/}{PurpleAir API}
}
\seealso{
\link{pas_downloadParseRawData}
}
