% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pas_filter.R
\name{pas_filter}
\alias{pas_filter}
\title{General purpose filtering for PurpleAir Synoptic objects}
\usage{
pas_filter(pas, ...)
}
\arguments{
\item{pas}{PurpleAir Synoptic *pas* object.}

\item{...}{Logical predicates defined in terms of the variables in the
`pas`. Multiple conditions are combined with & or seperated by a comma.
Only rows where the condition evaluates to TRUE are kept.}
}
\value{
A subset of the incoming *pas* object.
}
\description{
A generalized data filter for *pas* objects to
  choose rows/cases where conditions are true. Rows where the condition
  evaluates to NA are dropped.

  This function is essentially [dplyr::filter()]
  with an extra check on the validity of the *pas* object.
}
\examples{
library(AirSensor2)

nrow(example_pas_pm25)

# Washington
WA_pas <-
  example_pas_pm25 \%>\%
  pas_filter(stateCode == "WA")

nrow(WA_pas)

# Okanogan and Ferry Counties
Colville_Tribes_pas <-
  example_pas_pm25 \%>\%
  pas_filter(stateCode == "WA") \%>\%
  pas_filter(countyName \%in\% c("Okanogan", "Ferry"))

nrow(Colville_Tribes_pas)
}
\seealso{
[pas_filterArea()]

[pas_filterNear()]

[pas_filterNearMonitor()]
}
