% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pas_filterArea.R
\name{pas_filterArea}
\alias{pas_filterArea}
\title{Rectangle area filtering for PurpleAir Synoptic objects}
\usage{
pas_filterArea(pas = NULL, w = NULL, e = NULL, s = NULL, n = NULL)
}
\arguments{
\item{pas}{PurpleAir Synoptic \emph{pas} object.}

\item{w}{West edge of area bounding box (deg E).}

\item{e}{East edge of area bounding box (deg E).}

\item{s}{South edge of area bounding box (deg N).}

\item{n}{North edge of area bounding box (deg N).}
}
\value{
A subset of the given \emph{pas} object.
}
\description{
Filters \emph{pas} object sensors based on a bounding box.
}
\examples{
library(AirSensor2)

pas <- example_pas
range(pas$longitude)
range(pas$latitude)

Lane_County_pas <-
  pas \%>\%
  pas_filterArea(
    w = -124.16,
    e = -121.76,
    s = 43.43,
    n = 44.30
  )
range(Lane_County_pas$longitude)
range(Lane_County_pas$latitude)

if ( interactive() ) {
  ###pas_leaflet(Lane_County_pas)
}

}
\seealso{
\link{pas_filter}, \link{pas_filterNear}
}
