% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-pas.R
\name{pas_isEmpty}
\alias{pas_isEmpty}
\title{Test for an empty *pas* object}
\usage{
pas_isEmpty(pas = NULL)
}
\arguments{
\item{pas}{A *PurpleAir_synoptic* object.}
}
\value{
`TRUE` if no data exist in `pas`, `FALSE` otherwise.
}
\description{
Convenience function for `nrow(pas) == 0`.
This makes for more readable code in functions that need to test for this.
}
\examples{
library(AirSensor2)
example_pas_pm25 \%>\% pas_isEmpty() \%>\% print()
example_pas_pm25 \%>\% pas_filter(latitude > 90) \%>\% pas_isEmpty() \%>\% print()
}
