% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-plot.R
\name{pas_palette}
\alias{pas_palette}
\title{Color palettes for PurpleAir}
\usage{
pas_palette(pas = NULL, parameter = "pm2.5_60minutes", paletteName = NULL, ...)
}
\arguments{
\item{pas}{PurpleAir Synoptic \emph{pas} object.}

\item{parameter}{Value to generate colors for, e.g. \code{pm2.5_1hour}.}

\item{paletteName}{Optional name of an RColorBrewer paeltte, \emph{e.g.}
\code{"BuPu"} or \code{"Greens"} to use when a default paleltte is unavailable.}

\item{...}{Additional arguments passed on to \code{leaflet::color~} functions.}
}
\value{
An object that consists of a label and color dataframe, and
calculated color values from PurpleAir sensors
}
\description{
Generates color palettes for PurpleAir synoptic data with the
intention of having a reproducible functional color generator. Default palettes
are available for the following parameters:
\itemize{
\item{"pm2.5_~"}
\item{"humidity}
\item{"temperature}
}
}
