% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pas_pull.R
\name{pas_pull}
\alias{pas_pull}
\title{Returns a column of data from a PurpleAir synoptic object}
\usage{
pas_pull(pas, name = NULL)
}
\arguments{
\item{pas}{PurpleAir Synoptic \emph{pas} object.}

\item{name}{Name of the data column to return.}
}
\value{
A single column of data from the incoming \emph{pas} object.
}
\description{
A simple wrapper around \code{dplyr::pull()} to help with
readability of user recipes.
}
\examples{
library(AirSensor2)

# Lane (County) Regional Air Protection Agency
LRAPA_sensor_indices <-
  example_pas_pm25 \%>\%
  pas_filter(stringr::str_detect(name, "Ambassador")) \%>\%
  pas_pull("sensor_index")

print(LRAPA_sensor_indices)
}
