% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pat_applyCorrection.R
\name{pat_applyCorrection}
\alias{pat_applyCorrection}
\title{Apply correction to PurpleAir PM2.5 data}
\usage{
pat_applyCorrection(pat = NULL)
}
\arguments{
\item{pat}{Previously generated *hourly pat* object.}
}
\value{
A PurpleAir Timeseries *pat* object with an additional
`pm2.5_corrected` variable.
}
\description{
A correction equation is applied to fields of the incoming *pat* object to
generate a `pm2.5_corrected` time series which is added to the returned
*pat* object.
}
