% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-pat.R
\name{pat_check}
\alias{pat_check}
\title{Check *pat* object for validity}
\usage{
pat_check(pat)
}
\arguments{
\item{pat}{*pat* object.}
}
\value{
Returns `TRUE` invisibly if the *pat* object is valid.
}
\description{
Checks on the validity of an *pat* object. If any test
fails, this function will stop with a warning message.
}
\examples{
library(AirSensor2)

pat_check(example_pat)

# This would throw an error
if ( FALSE ) {

  broken_pat <- example_pat
  names(broken_pat) <- c('meta', 'bop')
  pat_check(broken_pat)

}

}
\seealso{
[pat_isValid()]
}
