% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pat_createNew.R
\name{pat_createNew}
\alias{pat_createNew}
\title{Create a new PurpleAir timeseries dataset.}
\usage{
pat_createNew(
  api_key = NULL,
  pas = NULL,
  sensor_index = NULL,
  startdate = NULL,
  enddate = NULL,
  timezone = "UTC",
  average = 0,
  fields = PurpleAir_HISTORY_PM25_FIELDS,
  parallel = FALSE,
  baseUrl = "https://api.purpleair.com/v1/sensors",
  verbose = FALSE
)
}
\arguments{
\item{api_key}{PurpleAir API Read Key. If \code{api_key = NULL}, it
will be obtained using \code{getAPIKey("PurpleAir-read")}.
See \code{MazamaCoreUtils::\link[MazamaCoreUtils:setAPIKey]{setAPIKey}}.}

\item{pas}{Previously generated \emph{pas} object containing \code{sensor_index}.}

\item{sensor_index}{PurpleAir sensor unique identifier.}

\item{startdate}{Desired start time (ISO 8601) or \code{POSIXct}.}

\item{enddate}{Desired end time (ISO 8601) or \code{POSIXct}.}

\item{timezone}{Olson timezone used to interpret dates.}

\item{average}{Temporal averaging in minutes performed by PurpleAir. One of:
0 (raw), 10, 30, 60 (hour), 360, 1440 (day).}

\item{fields}{Character string with PurpleAir field names for the Get Sensor Data API.}

\item{parallel}{Logical specifying whether to attempt simultaneous downloads
using \code{parallel::\link[parallel:mcparallel]{mcparallel}}. (Not available
on Windows.)}

\item{baseUrl}{Base URL for the PurpleAir API.}

\item{verbose}{Logical controlling the generation of warning and error messages.}
}
\value{
A PurpleAir Timeseries \emph{pat} object.
}
\description{
Create a \code{pat} object for a specific \code{sensor_index}.
}
\note{
Parallel processing using \code{parallel = TRUE} is not available on
Windows machines.
}
\examples{
\donttest{
# Fail gracefully if any resources are not available
try({

# AirSensor2 package
library(AirSensor2)

# Set user's PurpleAir_API_READ_KEY
source('global_vars.R')
setAPIKey("PurpleAir-read", PurpleAir_API_READ_KEY)

# Initialize spatial datasets
initializeMazamaSpatialUtils()

# Create a pat object
pat <-
  pat_createNew(
    pas = example_pas,
    sensor_index = "76545",
    startdate = "2023-01-01",
    enddate = "2023-01-08",
    timezone = "UTC",
    verbose = TRUE
  )

str(pat$meta)

# Run data download in parallel (faster)
pat <-
  pat_createNew(
    pas = example_pas,
    sensor_index = "76545",
    startdate = "2023-01-01",
    enddate = "2023-01-08",
    timezone = "UTC",
    parallel = TRUE,
    verbose = TRUE
  )

}, silent = FALSE)
}
}
\references{
\href{https://www2.purpleair.com}{PurpleAir}

\href{https://api.purpleair.com}{PurpleAir API}

\href{https://www2.purpleair.com/policies/terms-of-service}{PurpleAir Terms of service}

\href{https://www2.purpleair.com/pages/license}{PurpleAir Data license}

\href{https://www2.purpleair.com/pages/attribution}{PurpleAir Data Attribution}
}
