% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pat_filterDate.R
\name{pat_filterDate}
\alias{pat_filterDate}
\title{Date filtering for *pat* time series objects}
\usage{
pat_filterDate(
  pat = NULL,
  startdate = NULL,
  enddate = NULL,
  timezone = NULL,
  unit = "sec",
  ceilingStart = FALSE,
  ceilingEnd = FALSE
)
}
\arguments{
\item{pat}{*pat* object.}

\item{startdate}{Desired start datetime (ISO 8601).}

\item{enddate}{Desired end datetime (ISO 8601).}

\item{timezone}{Olson timezone used to interpret dates.}

\item{unit}{Units used to determine time at end-of-day.}

\item{ceilingStart}{Logical instruction to apply
[lubridate::ceiling_date()] to the `startdate` rather than
[lubridate::floor_date()]}

\item{ceilingEnd}{Logical instruction to apply
[lubridate::ceiling_date()] to the `enddate` rather than
[lubridate::floor_date()]}
}
\value{
A subset of the incoming *pat* time series object.
(A list with `meta` and `data` dataframes.)
}
\description{
Subsets a *pat* object by date. This function
always filters to day-boundaries. For sub-day filtering, use
`pat_filterDatetime()`.

Dates can be anything that is understood by `MazamaCoreUtils::parseDatetime()`
including either of the following recommended formats:

\itemize{
\item{`"YYYYmmdd"`}
\item{`"YYYY-mm-dd"`}
}

Timezone determination precedence assumes that if you are passing in
`POSIXct` values then you know what you are doing.

\enumerate{
\item{get timezone from `startdate` if it is `POSIXct`}
\item{use passed in `timezone`}
\item{get timezone from `pat`}
}
}
\note{
The returned data will run from the beginning of `startdate` until
the **beginning** of `enddate` -- *i.e.* no values associated
with `enddate` will be returned. The exception being when
`enddate` is less than 24 hours after `startdate`. In that case, a
single day is returned.
}
\seealso{
[pat_filter()]

[pat_filterDatetime()]
}
