% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pat_filterDatetime.R
\name{pat_filterDatetime}
\alias{pat_filterDatetime}
\title{Datetime filtering for *pat* time series objects}
\usage{
pat_filterDatetime(
  pat = NULL,
  startdate = NULL,
  enddate = NULL,
  timezone = NULL,
  unit = "sec",
  ceilingStart = FALSE,
  ceilingEnd = FALSE
)
}
\arguments{
\item{pat}{*pat* object.}

\item{startdate}{Desired start datetime (ISO 8601).}

\item{enddate}{Desired end datetime (ISO 8601).}

\item{timezone}{Olson timezone used to interpret dates.}

\item{unit}{Units used to determine time at end-of-day.}

\item{ceilingStart}{Logical instruction to apply
[lubridate::ceiling_date()] to the `startdate` rather than
[lubridate::floor_date()]}

\item{ceilingEnd}{Logical instruction to apply
[lubridate::ceiling_date()] to the `enddate` rather than
[lubridate::floor_date()]}
}
\value{
A subset of the incoming *pat* time series object.
(A list with `meta` and `data` dataframes.)
}
\description{
Subsets a *pat* object by datetime. This function
allows for sub-day filtering as opposed to `pat_filterDate()` which
always filters to day-boundaries.

Datetimes can be anything that is understood by
`MazamaCoreUtils::parseDatetime()`. For non-`POSIXct` values,
the recommended format is `"YYYY-mm-dd HH:MM:SS"`.

Timezone determination precedence assumes that if you are passing in
`POSIXct` values then you know what you are doing.

\enumerate{
\item{get timezone from `startdate` if it is `POSIXct`}
\item{use passed in `timezone`}
\item{get timezone from `pat`}
}
}
\seealso{
[pat_filter()]

[pat_filterDate()]
}
