% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-pat.R
\name{pat_getDataFrame}
\alias{pat_getDataFrame}
\alias{pat_getData}
\alias{pat_getMeta}
\title{Extract dataframes from *pat* objects}
\usage{
pat_getData(pat)

pat_getMeta(pat)
}
\arguments{
\item{pat}{*pat* object to extract dataframe from.}
}
\value{
A dataframe from the *pat* object.
}
\description{
These functions are convenient wrappers for extracting the dataframes that
comprise a *pat* object. These functions are designed to be useful when
manipulating data in a pipeline using `\%>\%`.

Below is a table showing equivalent operations for each function.

`pat_getData(pat)` is equivalent to `pat$data`.

`pat_getMeta(pat)` is equivalent to `pat$meta`.
}
