% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-pat.R
\name{pat_isEmpty}
\alias{pat_isEmpty}
\title{Test for empty *pat* object}
\usage{
pat_isEmpty(pat)
}
\arguments{
\item{pat}{*pat* object}
}
\value{
`TRUE` if no data exist in `pat`, `FALSE` otherwise.
}
\description{
Convenience function for `nrow(pat$data) == 0`.
This makes for more readable code in functions that need to test for this.
}
\examples{
library(AirSensor2)

pat_isEmpty(example_pat)

}
