% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-pat.R
\name{pat_isValid}
\alias{pat_isValid}
\title{Test *pat* object for correct structure}
\usage{
pat_isValid(pat = NULL, verbose = FALSE)
}
\arguments{
\item{pat}{*pat* object}

\item{verbose}{Logical specifying whether to produce detailed warning messages.}
}
\value{
`TRUE` if `pat` has the correct structure,
`FALSE` otherwise.
}
\description{
The `pat` is checked for the presence of core
`meta` and `data` columns.

Core `meta` columns include:

\itemize{
  \item{`deviceDeploymentID` -- unique identifier (see \pkg{MazmaLocationUtils})}
  \item{`deviceID` -- device identifier}
  \item{`locationID` -- location identifier (see \pkg{MazmaLocationUtils})}
  \item{`locationName` -- English language name}
  \item{`longitude` -- decimal degrees E}
  \item{`latitude` -- decimal degrees N}
  \item{`elevation` -- elevation of station in m}
  \item{`countryCode` -- ISO 3166-1 alpha-2}
  \item{`stateCode` -- ISO 3166-2 alpha-2}
  \item{`timezone` -- Olson time zone}
}

Core `data` columns include:

\itemize{
  \item{`datetime` -- measurement time (UTC)}
}
}
\examples{
library(AirSensor2)

pat_isValid(example_pat)

}
