% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pat_toMonitor.R
\name{pat_toMonitor}
\alias{pat_toMonitor}
\title{Convert a PurpleAir Timeseries to a 'mts_monitor' object}
\usage{
pat_toMonitor(pat = NULL, applyCorrection = TRUE)
}
\arguments{
\item{pat}{Previously generated *pat* object.}

\item{applyCorrection}{Logical specifying whether to apply the EPA
correction algorithm. (See [pat_applyCorrection()].)}
}
\value{
An AirMonitor package *mts_monitor* object.
}
\description{
Enhance an hourly PurpleAir Timeseries *pat* object created
with `pat_createHourly()` and create an object of class `mts_monitor`
for use with the [AirMonitor](https://mazamascience.github.io/AirMonitor/) package.
}
\examples{
\donttest{
# Fail gracefully if any resources are not available
try({

# AirSensor2 package
library(AirSensor2)
monitor <-
  example_pat_epa \%>\%
  pat_toMonitor()

AirMonitor::monitor_timeseriesPlot(monitor, shadedNight = TRUE)

}, silent = FALSE)
}
}
\references{
[PurpleAir](https://www2.purpleair.com)

[PurpleAir API](https://api.purpleair.com)

[PurpleAir Terms of service](https://www2.purpleair.com/policies/terms-of-service)

[PurpleAir Data license](https://www2.purpleair.com/pages/license)

[PurpleAir Data Attribution](https://www2.purpleair.com/pages/attribution)
}
