% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PurpleAir_API.R
\name{PurpleAir_checkAPIKey}
\alias{PurpleAir_checkAPIKey}
\title{Check the validity and type for the provided \code{api_key}}
\usage{
PurpleAir_checkAPIKey(
  api_key = NULL,
  baseUrl = "https://api.purpleair.com/v1/keys"
)
}
\arguments{
\item{api_key}{PurpleAir API key.}

\item{baseUrl}{URL endpoint for the "Check Key" API.}
}
\value{
List containing key type information.
}
\description{
Sends a request to the PurpleAirAPI API endpoint described at:
\url{https://api.purpleair.com/#api-keys-check-api-key}
}
\examples{
\donttest{
# Fail gracefully if any resources are not available
try({

library(AirSensor2)

  PurpleAir_checkAPIKey(
    api_key = PurpleAir_API_READ_KEY
  )

}, silent = FALSE)
}
}
