% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PurpleAir_API.R
\name{PurpleAir_createGroup}
\alias{PurpleAir_createGroup}
\title{Create a new group.}
\usage{
PurpleAir_createGroup(
  api_key = NULL,
  name = NULL,
  baseUrl = "https://api.purpleair.com/v1/groups"
)
}
\arguments{
\item{api_key}{PurpleAir API WRITE key.}

\item{name}{Human readable name associated with the new group.}

\item{baseUrl}{URL endpoint for the "Create Group" API.}
}
\value{
List containing all members of the specified group.
}
\description{
Sends a request to the PurpleAirAPI API endpoint described at:
\url{https://api.purpleair.com/#api-groups-create-group}
}
\examples{
\donttest{
# Fail gracefully if any resources are not available
try({

library(AirSensor2)

  PurpleAir_createGroup(
    api_key = PurpleAir_API_WRITE_KEY,
    name = "My new group"
  )

}, silent = FALSE)
}
}
