% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PurpleAir_createMonitor.R
\name{PurpleAir_createMonitor}
\alias{PurpleAir_createMonitor}
\title{Create a new PurpleAir 'mts_monitor' object}
\usage{
PurpleAir_createMonitor(
  api_key = NULL,
  pas = NULL,
  sensor_index = NULL,
  parameter = c("pm2.5_atm"),
  startdate = NULL,
  enddate = NULL,
  timezone = "UTC",
  verbose = FALSE
)
}
\arguments{
\item{api_key}{PurpleAir API READ Key. If \code{api_key = NULL}, it
will be obtained using \code{getAPIKey("PurpleAir-read")}.}

\item{pas}{Previously generated \emph{pas} object containing \code{sensor_index}.}

\item{sensor_index}{PurpleAir sensor identifier.}

\item{parameter}{Parameter to use for data ("pm2.5_atm")}

\item{startdate}{Desired start time (ISO 8601) or \code{POSIXct}.}

\item{enddate}{Desired end time (ISO 8601) or \code{POSIXct}.}

\item{timezone}{Olson timezone used to interpret dates.}

\item{verbose}{Logical controlling the generation of warning and error messages.}
}
\value{
An AirMonitor package \emph{mts_monitor} object.
}
\description{
Download, parse and enhance hourly timeseries data from PurpleAir
and create an object of class \code{mts_monitor} for use with the AirMonitor
package.
}
\note{
This is a very earily implementation that does not include any QC or
correction equations.
}
\examples{
\donttest{
# Fail gracefully if any resources are not available
try({

library(AirSensor2)

initializeMazamaSpatialUtils()

mon <-
  PurpleAir_createMonitor(
    api_key = PurpleAir_API_READ_KEY,
    pas = MY_PAS,
    sensor_index = "76545",
    startdate = "2023-01-01",
    enddate = "2023-01-08",
    timezone = "UTC",
    verbose = TRUE
  )

}, silent = FALSE)
}
}
\references{
\href{https://www2.purpleair.com}{PurpleAir}

\href{https://api.purpleair.com}{PurpleAir API}

\href{https://www2.purpleair.com/policies/terms-of-service}{PurpleAir Terms of service}

\href{https://www2.purpleair.com/pages/license}{PurpleAir Data license}

\href{https://www2.purpleair.com/pages/attribution}{PurpleAir Data Attribution}
}
