% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PurpleAir_API.R
\docType{data}
\name{PurpleAir_HISTORY_HOURLY_PM25_FIELDS}
\alias{PurpleAir_HISTORY_HOURLY_PM25_FIELDS}
\title{Comma-separated list of fields needed for simple PM2.5 data analysis.}
\format{
String with comma-separated field names
}
\usage{
PurpleAir_HISTORY_HOURLY_PM25_FIELDS
}
\description{
Character string with default PurpleAir field names used in
\code{pat_downloadParaseRawData()}. These fields are sufficient for most
QC algorithms and include most of the "information and status" fields,
"humidity", "temperature", and the PM2.5 fields for outdoor sensors.
}
\references{
\href{https://api.purpleair.com/#api-sensors-get-sensor-history-csv}{Get Sensor History API}
}
\keyword{datasets}
