% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PurpleAir_API.R
\docType{data}
\name{PurpleAir_PAT_EPA_HOURLY_FIELDS}
\alias{PurpleAir_PAT_EPA_HOURLY_FIELDS}
\title{Comma-separated list of fields needed for simple PM2.5 data analysis}
\format{
String with comma-separated field names
}
\usage{
PurpleAir_PAT_EPA_HOURLY_FIELDS
}
\description{
Character string with a minimal set of  PurpleAir field names
used in \code{pat_downloadParaseRawData()}. These fields are sufficient for
basic QC and correction algorithms.

Included fields:
\preformatted{
[1] "humidity"     "temperature"  "pm2.5_atm"    "pm2.5_atm_a"
[5] "pm2.5_atm_b"  "pm2.5_cf_1"   "pm2.5_cf_1_a" "pm2.5_cf_1_b"
}
}
\references{
\href{https://api.purpleair.com/#api-sensors-get-sensor-history-csv}{Get Sensor History API}
}
\keyword{datasets}
