% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PurpleAir_API.R
\docType{data}
\name{PurpleAir_SENSOR_METADATA_FIELDS}
\alias{PurpleAir_SENSOR_METADATA_FIELDS}
\title{Comma-separated list of metadata fields used to create a \emph{pas} object}
\format{
String with comma-separated field names
}
\usage{
PurpleAir_SENSOR_METADATA_FIELDS
}
\description{
Character string with PurpleAir field names used in
\code{pas_downloadParseRawData()}. These fields exclude all measurements
but retain a minimal set of metadata useful for filtering. This \emph{pas} object
can then be passed on to \emph{pat} creation functions.

A metadata-only \emph{pas} object can be useful when searching for historical
data using \code{\link{pas_filterDate}}.

Included fields:
\itemize{
\item{\code{name}}
\item{\code{model}}
\item{\code{hardware}}
\item{\code{location_type}}
\item{\code{private}}
\item{\code{latitude}}
\item{\code{longitude}}
\item{\code{altitude}}
\item{\code{position_rating}}
\item{\code{firmware_version}}
\item{\code{firmware_upgrade}}
\item{\code{uptime}}
\item{\code{last_seen}}
\item{\code{last_modified}}
\item{\code{date_created}}
}
}
\references{
\href{https://api.purpleair.com/#api-sensors-get-sensor-data}{Get Sensor Data API}
}
\keyword{datasets}
