% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initializeMazamaSpatialUtils.R
\name{spatialIsInitialized}
\alias{spatialIsInitialized}
\title{Check whether MazamaSpatialUtils has been initialized}
\usage{
spatialIsInitialized()
}
\value{
Logical `TRUE` if initialization has been performed in the current
  session, `FALSE` otherwise.
}
\description{
Convenience function that checks whether
`initializeMazamaSpatialUtils()` has been run in the current R session.
}
