% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/syn_leaflet.R
\name{syn_leaflet}
\alias{syn_leaflet}
\title{Leaflet interactive map sensor metadata}
\usage{
syn_leaflet(
  syn = NULL,
  parameter = "pm25",
  paletteName = NULL,
  radius = 10,
  opacity = 0.8,
  maptype = "terrain"
)
}
\arguments{
\item{syn}{Synoptic metadata \emph{~_synoptic} object.}

\item{parameter}{Value to plot, e.g. \code{pm25}.}

\item{paletteName}{\pkg{RColorBrewer} palette name to use when \code{parameter}
is something other than:
\itemize{
\item{"pm2.5_~"}
\item{"humidity}
\item{"temperature}
}}

\item{radius}{Radius (pixels) of monitor circles.}

\item{opacity}{Opacity of monitor circles.}

\item{maptype}{Optional name of leaflet ProviderTiles to use, e.g. \code{terrain}.}
}
\value{
A leaflet "plot" object which, if not assigned, is rendered in
Rstudio's 'Viewer' tab.
}
\description{
This function creates interactive maps that will be displayed in
RStudio's 'Viewer' tab.

Typical usage would be to use the \code{parameter} argument to display pm25
values from one of:
\itemize{
\item{"pm2.5_10minute"}
\item{"pm2.5_30minute"}
\item{"pm2.5_60minute"}
\item{"pm2.5_6hour"}
\item{"pm2.5_24hour"}
\item{"pm2.5_1week"}
}
}
\details{
The \code{maptype} argument is mapped onto leaflet "ProviderTile"
names. Current mappings include:
\enumerate{
\item{"roadmap"}{ -- "OpenStreetMap"}
\item{"satellite"}{ -- "Esri.WorldImagery"}
\item{"terrain"}{ -- "Esri.WorldTopoMap"}
\item{"toner"}{ -- "Stamen.Toner"}
}

If a character string not listed above is provided, it will be used as the
underlying map tile if available. See
\url{https://leaflet-extras.github.io/leaflet-providers/} for a list of
"provider tiles" to use as the background map.
}
\note{
The \code{paletteName} parameter can take the name of an RColorBrewer
paeltte, \emph{e.g.} \code{"BuPu"} or \code{"Greens"}.
}
\examples{
library(AirSensor2)

if ( interactive() ) {
  syn_leaflet(example_syn, parameter = "pm2.5_60minute")

  syn_leaflet(example_syn, parameter = "temperature")

  syn_leaflet(example_syn, parameter = "humidity")
}
}
