% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-plot.R
\name{syn_palette}
\alias{syn_palette}
\title{Color palettes for Synoptic data}
\usage{
syn_palette(syn = NULL, parameter = "pm2.5", paletteName = NULL, ...)
}
\arguments{
\item{syn}{PurpleAir Synoptic \emph{syn} object.}

\item{parameter}{Value to generate colors for, e.g. \code{pm2.5}.}

\item{paletteName}{Optional name of an RColorBrewer paeltte, \emph{e.g.}
\code{"BuPu"} or \code{"Greens"} to use when a default paleltte is unavailable.}

\item{...}{Additional arguments passed on to \code{leaflet::color~} functions.}
}
\value{
An object that consists of a label and color dataframe, and
calculated color values.
}
\description{
Generates color palettes for PurpleAir synoptic data with the
intention of having a reproducible functional color generator. Default palettes
are available for the following parameters:
\itemize{
\item{"pm2.5"}
\item{"humidity}
\item{"temperature}
}
}
