% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/synoptic_leaflet.R
\name{synoptic_leaflet}
\alias{synoptic_leaflet}
\title{Leaflet interactive map sensor metadata}
\usage{
synoptic_leaflet(
  synoptic = NULL,
  parameter = "pm2.5",
  paletteName = NULL,
  radius = 10,
  opacity = 0.8,
  maptype = "terrain",
  extraVars = NULL
)
}
\arguments{
\item{synoptic}{Synoptic metadata *~_synoptic* object.}

\item{parameter}{Value to plot, e.g. `pm25`.}

\item{paletteName}{\pkg{RColorBrewer} palette name to use when `parameter`
is something other than:
\itemize{
\item{"pm2.5"}
\item{"humidity}
\item{"temperature}
}}

\item{radius}{Radius (pixels) of monitor circles.}

\item{opacity}{Opacity of monitor circles.}

\item{maptype}{Optional name of leaflet ProviderTiles to use, e.g. `terrain`.}

\item{extraVars}{Character vector of additional column names to be shown in
leaflet popups.}
}
\value{
A leaflet "plot" object which, if not assigned, is rendered in
Rstudio's 'Viewer' tab.
}
\description{
This function creates interactive maps that will be displayed in
RStudio's 'Viewer' tab.

Typical usage would be to use the `parameter` argument to display pm25
values:
\itemize{
\item{"pm2.5"}
}
}
\details{
The `maptype` argument is mapped onto leaflet "ProviderTile"
names. Current mappings include:
\describe{
\item{"roadmap"}{ -- "OpenStreetMap"}
\item{"satellite"}{ -- "Esri.WorldImagery"}
\item{"terrain"}{ -- "Esri.WorldTopoMap"}
\item{"toner"}{ -- "Stamen.Toner"}
}

If a character string not listed above is provided, it will be used as the
underlying map tile if available. See
\url{https://leaflet-extras.github.io/leaflet-providers/} for a list of
"provider tiles" to use as the background map.
}
\note{
The `paletteName` parameter can take the name of an RColorBrewer
paeltte, *e.g.* `"BuPu"` or `"Greens"`.
}
