% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-plot.R
\name{synoptic_palette}
\alias{synoptic_palette}
\title{Color palettes for Synoptic data}
\usage{
synoptic_palette(synoptic = NULL, parameter = "pm2.5", paletteName = NULL, ...)
}
\arguments{
\item{synoptic}{PurpleAir Synoptic *synoptic* object.}

\item{parameter}{Value to generate colors for, e.g. `pm2.5`.}

\item{paletteName}{Optional name of an RColorBrewer paeltte, *e.g.*
`"BuPu"` or `"Greens"` to use when a default palette is unavailable.}

\item{...}{Additional arguments passed on to `leaflet::color~` functions.}
}
\value{
A list containing: 1) a label and color dataframe; and 2)
a vector of color values calculated from the data found in `synoptic`.
}
\description{
Generates color palettes for PurpleAir synoptic data with the
intention of having a reproducible functional color generator. Default palettes
are available for the following parameters:
\itemize{
\item{"pm2.5"}
\item{"humidity}
\item{"temperature}
}
}
