% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Clarity_API.R
\name{Clarity_getAllOpenHourly}
\alias{Clarity_getAllOpenHourly}
\title{Retrieve current hourly data for all Open Data sensors.}
\usage{
Clarity_getAllOpenHourly(
  api_key = NULL,
  format = "USFS",
  baseUrl =
    "https://clarity-data-api.clarity.io/v1/open/all-recent-measurement/pm25/hourly"
)
}
\arguments{
\item{api_key}{Clarity API READ key.}

\item{format}{Customized output format (currently only "USFS").}

\item{baseUrl}{URL endpoint.}
}
\value{
List containing four data frames: \code{synoptic}, \code{QCFlag},
\code{pm2.5} and \code{nowcast}.
}
\description{
Sends a request to the Clarity API endpoint for Open Data.

\itemize{
  \item{\emph{Measurement data from} -- All open datasources}
  \item{\emph{Measurements returned} -- Hourly values for last 3 hours}
  \item{\emph{PM2.5 Mass Concentration} -- 1-Hour Mean Nowcast}
}
}
\examples{
\donttest{
# Fail gracefully if any resources are not available
try({

library(AirSensor2)

  Clarity_getAllOpenHourly(
    api_key = Clarity_API_READ_KEY,
    format = "USFS"
  )

}, silent = FALSE)
}
}
