% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Clarity_API.R
\name{Clarity_getAllOpenIndividual}
\alias{Clarity_getAllOpenIndividual}
\title{Retrieve current individual records for all Open Data sensors.}
\usage{
Clarity_getAllOpenIndividual(
  api_key = NULL,
  format = c("USFS2", "USFS"),
  baseUrl =
    "https://clarity-data-api.clarity.io/v1/open/all-recent-measurement/pm25/individual"
)
}
\arguments{
\item{api_key}{Clarity API READ Key. If \code{api_key = NULL}, it
will be obtained using \code{getAPIKey("Clarity-read")}.}

\item{format}{Customized output format ("USFS2", "USFS").}

\item{baseUrl}{URL endpoint.}
}
\value{
List containing five data frames: \code{meta}, \code{pm2.5_QCFlag},
\code{pm2.5}, \code{nowcast_QCFlag} and \code{nowcast}.
}
\description{
Sends a request to the Clarity API endpoint for Open Data.

When \code{format = "USFS2"}, two additional fields are returned in the
"meta" dataframe: \code{calibrationId} and \code{calibrationCategory}.

\itemize{
  \item{\emph{Measurement data from} -- All open datasources}
  \item{\emph{Measurements returned} -- Individual values for last complete
  hour (hour-aligned) plus fraction of current hour}
  \item{\emph{PM2.5 Mass Concentration} -- Individual sample}
}
}
\examples{
\donttest{
# Fail gracefully if any resources are not available
try({

library(AirSensor2)

  Clarity_getAllIndividualOpen(
    api_key = Clarity_API_READ_KEY,
    format = "USFS2"
  )

}, silent = FALSE)
}
}
