% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Clarity_API.R
\name{Clarity_getOpenIndividual}
\alias{Clarity_getOpenIndividual}
\title{Retrieve current individual records from a single source.}
\usage{
Clarity_getOpenIndividual(
  api_key = NULL,
  datasourceId = NULL,
  format = c("USFS"),
  baseUrl = "https://clarity-data-api.clarity.io/v1/open/datasource-measurement"
)
}
\arguments{
\item{api_key}{Clarity API READ key.}

\item{datasourceId}{Clarity sensor identifier.}

\item{format}{Customized output format (currently only "USFS").}

\item{baseUrl}{URL endpoint.}
}
\value{
List containing four data frames: \code{meta}, \code{QCFlag},
\code{pm2.5} and \code{nowcast}.
}
\description{
Sends a request to the Clarity API endpoint for Open Data.

\itemize{
  \item{\emph{Measurement data from} -- Drill down on one open datasource}
  \item{\emph{Measurements returned} -- Individual values for last 3 days}
  \item{\emph{PM2.5 Mass Concentration} -- Individual sample}
}
}
\examples{
\donttest{
# Fail gracefully if any resources are not available
try({

library(AirSensor2)

  Clarity_getOpenIndividual(
    api_key = Clarity_API_READ_KEY,
    datasourceId = "DAABL1560",
    format = "USFS"
  )

}, silent = FALSE)
}
}
