% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\encoding{UTF-8}
\name{example_pat}
\alias{example_pat}
\title{Example enhanced Purple Air Synoptic dataset}
\format{
A tibble with 1948 rows and 50 columns of data.
}
\source{
https://www2.purpleair.com
}
\usage{
example_pat
}
\description{
The \code{example_pas} dataset provides a quickly loadable
version of a \emph{pas} object for practicing and code examples.
This dataset contains data for sensors in Washington and Oregon
and was generated on 2023-02-15 by running:

\preformatted{
library(AirSensor2)

example_pat <-
  pat_createNew(
    api_key = PURPLE_AIR_API_READ_KEY,
    pas = example_pas,
    sensor_index = "76545",
    startdate = "2023-01-01",
    enddate = "2023-01-02",
    timezone = "America/Los_Angeles"
  )

save(example_pat, file = "data/example_pat.rda")
}
}
\seealso{
example_pas
}
\keyword{datasets}
