% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\encoding{UTF-8}
\name{example_pat}
\alias{example_pat}
\title{Example PurpleAir Timeseries dataset}
\format{
A list containing two tibbles named 'meta' and 'data'.
}
\source{
https://www2.purpleair.com
}
\usage{
example_pat
}
\description{
The \code{example_pat} dataset provides a quickly loadable
version of a \emph{pat} object for practicing and code examples.
This dataset contains 24 hours worth of data for a single PurpleAir sensor
and was generated on 2024-03-12 by running:

\preformatted{
library(AirSensor2)

source("global_vars.R") # contains PurpleAir_API_READ_KEY

example_pat <-
  pat_create(
    api_key = PurpleAir_API_READ_KEY,
    pas = example_pas_metadata,
    sensor_index = "95189",
    startdate = "2024-03-01",
    enddate = "2024-03-03",
    timezone = "America/Los_Angeles",
    average = 0,
    fields = PurpleAir_PAT_QC_FIELDS
  )

save(example_pat, file = "data/example_pat.rda")
#' }
}
\seealso{
example_pas
}
\keyword{datasets}
