% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pas_downloadParseRawData.R
\name{pas_downloadParseRawData}
\alias{pas_downloadParseRawData}
\title{Download synoptic data from PurpleAir with newer API}
\usage{
pas_downloadParseRawData(
  apiReadKey = NULL,
  maxAge = 604800,
  outsideOnly = TRUE,
  west = NULL,
  east = NULL,
  south = NULL,
  north = NULL,
  baseUrl = "https://api.purpleair.com/v1/sensors"
)
}
\arguments{
\item{apiReadKey}{PurpleAir API Read Key. If \code{apiReadKey = NULL}, it
will be obtained using \code{getAPIKey("PurpleAir-read")}.
See \code{MazamaCoreUtils::\link[MazamaCoreUtils:setAPIKey]{setAPIKey}}.}

\item{maxAge}{Filter results to only include sensors modified or updated
within the last number of seconds. Using a value of 0 will match sensors of any age.}

\item{outsideOnly}{Logical specifying whether to restrict requests to outside sensors only.}

\item{west}{Longitude of the western edge of the bounding box in which to find sensors.}

\item{east}{Longitude of the eastern edge of the bounding box in which to find sensors.}

\item{south}{Latitude of the southern edge of the bounding box in which to find sensors.}

\item{north}{Latitude of the northern edge of the bounding box in which to find sensors.}

\item{baseUrl}{Base URL for the PurpleAir API.}
}
\value{
Dataframe of synoptic PurpleAir data.
}
\description{
Download and parse synoptic data for the PurpleAir network
of particulate sensors.

The synoptic data provides a view of the entire PurpleAir network and
includes both metadata and recent PM2.5 averages for each deployed sensor.
}
\examples{
\donttest{
# Fail gracefully if any resources are not available
try({

library(AirSensor2)

pas_raw <-
  pas_downloadParseRawData(
    apiReadKey = NULL,
    maxAge = 3600 * 24,
    outsideOnly = TRUE,
    west = -125,
    east = -117,
    south = 42,
    north = 49
  )

View(pas_raw[1:100,])

}, silent = FALSE)
}
}
\references{
\href{https://api.purpleair.com}{PurpleAir API}
}
