% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pas_filterDate.R
\name{pas_filterDate}
\alias{pas_filterDate}
\title{Date filtering for PurpleAir Synoptic objects}
\usage{
pas_filterDate(pas = NULL, startdate = NULL, enddate = NULL, timezone = NULL)
}
\arguments{
\item{pas}{PurpleAir Synoptic \emph{pas} object.}

\item{startdate}{Desired start datetime (ISO 8601).}

\item{enddate}{Desired end datetime (ISO 8601).}

\item{timezone}{Optional Olson timezone used to interpret dates.}
}
\description{
Subsets a \emph{pas} object by date.

Dates can be anything that is understood by \code{MazamaCoreUtils::parseDatetime()}
including either of the following recommended formats:

\itemize{
\item{\code{"YYYYmmdd"}}
\item{\code{"YYYY-mm-dd"}}
}

Timezone determination precedence assumes that if you are passing in
\code{POSIXct} values then you know what you are doing.

\enumerate{
\item{get timezone from \code{startdate} if it is \code{POSIXct}}
\item{use passed in \code{timezone}}
}
}
\note{
The incoming \code{pas} object must contain the \code{date_created} and
\code{last_modified} fields. The returned \emph{pas} object will retain sites
where \code{startdate <= last_seen && enddate >=  pas$date_created}.
}
\examples{
library(AirSensor2)

august_2018 <-
 example_pas_historical \%>\%
  pas_filterDate(
    startdate = 20180701,
    enddate = 20181101,
    timezone = "America/Los_Angeles"
 )

fields <- c("sensor_index", "name", "date_created", "last_seen")
head(august_2018[,fields])

}
\seealso{
\link{pas_filter}
}
