% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-pas.R
\name{pas_hasSpatial}
\alias{pas_hasSpatial}
\title{Test for spatial metadata in \emph{pas} object}
\usage{
pas_hasSpatial(pas = NULL)
}
\arguments{
\item{pas}{A PurpleAir Synoptic \emph{pas} object.}
}
\value{
\code{TRUE} if \code{pas} contains core spatial metadata,
\code{FALSE} otherwise.
}
\description{
Tests for the existence of the following core spatial metadata
columns:

\itemize{
  \item{longitude -- decimal degrees E}
  \item{latitude -- decimal degrees N}
  \item{timezone -- Olson timezone}
  \item{countryCode -- ISO 3166-1 alpha-2}
  \item{stateCode -- ISO 3166-2 alpha-2}
}
}
\examples{
library(AirSensor2)
example_pas_pm25_raw \%>\% pas_hasSpatial() \%>\% print()
example_pas_pm25 \%>\% pas_hasSpatial() \%>\% print()
}
