% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-pas.R
\name{pas_isEmpty}
\alias{pas_isEmpty}
\title{Test for an empty \emph{pas} object}
\usage{
pas_isEmpty(pas = NULL)
}
\arguments{
\item{pas}{A \emph{pa_synoptic} object.}
}
\value{
\code{TRUE} if no data exist in \code{pas}, \code{FALSE} otherwise.
}
\description{
Convenience function for \code{nrow(pas) == 0}.
This makes for more readable code in functions that need to test for this.
}
\examples{
library(AirSensor2)
example_pas \%>\% pas_isEmpty() \%>\% print()
example_pas \%>\% pas_filter(latitude > 90) \%>\% pas_isEmpty() \%>\% print()
}
