% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pat_filter.R
\name{pat_filter}
\alias{pat_filter}
\title{General purpose data filtering for \emph{pat} time series objects}
\usage{
pat_filter(pat, ...)
}
\arguments{
\item{pat}{\emph{pat} object.}

\item{...}{Logical predicates defined in terms of the variables in
\code{pat$data}.}
}
\value{
A subset of the incoming \code{pat} time series object.
(A list with \code{meta} and \code{data} dataframes.)
}
\description{
A generalized data filter for \emph{pat} objects to
choose rows/cases where conditions are true.  Multiple conditions are
combined with \code{&} or separated by a comma. Only rows where the condition
evaluates to TRUE are kept. Rows where the condition evaluates to \code{NA}
are dropped.

If an empty \emph{pat} object is passed in, it is immediately returned,
allowing for multiple filtering steps to be piped together and only checking
for an empty \emph{pat} object at the end of the pipeline.
}
\note{
Filtering is done on values in \code{pat$data}.
}
\examples{

library(AirSensor2)

# Unhealthy values
example_pat \%>\%
  pat_filter(pm2.5_atm_a > 55.5, pm2.5_atm_b > 55.5) \%>\%
  pat_getData() \%>\%
  head()

}
\seealso{
\link{pat_filterDate}

\link{pat_filterDatetime}
}
