% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-pat.R
\name{pat_isValid}
\alias{pat_isValid}
\title{Test \emph{pat} object for correct structure}
\usage{
pat_isValid(pat = NULL, verbose = FALSE)
}
\arguments{
\item{pat}{\emph{pat} object}

\item{verbose}{Logical specifying whether to produce detailed warning messages.}
}
\value{
\code{TRUE} if \code{pat} has the correct structure,
\code{FALSE} otherwise.
}
\description{
The \code{pat} is checked for the presence of core
\code{meta} and \code{data} columns.

Core \code{meta} columns include:

\itemize{
  \item{\code{deviceDeploymentID} -- unique identifier (see \pkg{MazmaLocationUtils})}
  \item{\code{deviceID} -- device identifier}
  \item{\code{locationID} -- location identifier (see \pkg{MazmaLocationUtils})}
  \item{\code{locationName} -- English language name}
  \item{\code{longitude} -- decimal degrees E}
  \item{\code{latitude} -- decimal degrees N}
  \item{\code{elevation} -- elevation of station in m}
  \item{\code{countryCode} -- ISO 3166-1 alpha-2}
  \item{\code{stateCode} -- ISO 3166-2 alpha-2}
  \item{\code{timezone} -- Olson time zone}
}

Core \code{data} columns include:

\itemize{
  \item{\code{datetime} -- measurement time (UTC)}
}
}
\examples{
library(AirSensor2)

pat_isValid(example_pat)

}
