% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pat_toMonitor.R
\name{pat_toMonitor}
\alias{pat_toMonitor}
\title{Convert a PurpleAir Timeseries to a 'mts_monitor' object}
\usage{
pat_toMonitor(pat = NULL)
}
\arguments{
\item{pat}{Previously generated \emph{pat} object.}
}
\value{
An AirMonitor package \emph{mts_monitor} object.
}
\description{
Enhance an hourly PurpleAir Timeseries \emph{pat} object created
with \code{pat_createHourly()} and create an object of class \code{mts_monitor}
for use with the \href{https://mazamascience.github.io/AirMonitor/}{AirMonitor} package.
}
\examples{
\donttest{
# Fail gracefully if any resources are not available
try({

# AirSensor2 package
library(AirSensor2)

monitor <-
  example_pat \%>\%
  pat_toMonitor(
  )

AirMonitor::monitor_timeseriesPlot(monitor, shadedNight = TRUE)
}, silent = FALSE)
}
}
\references{
\href{https://www2.purpleair.com}{PurpleAir}

\href{https://api.purpleair.com}{PurpleAir API}

\href{https://www2.purpleair.com/policies/terms-of-service}{PurpleAir Terms of service}

\href{https://www2.purpleair.com/pages/license}{PurpleAir Data license}

\href{https://www2.purpleair.com/pages/attribution}{PurpleAir Data Attribution}
}
