% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PurpleAir_createMonitor.R
\name{PurpleAir_createMonitor}
\alias{PurpleAir_createMonitor}
\title{Create a new PurpleAir 'mts_monitor' object}
\usage{
PurpleAir_createMonitor(
  api_key = NULL,
  pas = NULL,
  sensor_index = NULL,
  startdate = NULL,
  enddate = NULL,
  timezone = "UTC",
  correction_FUN = PurpleAir_correction,
  verbose = FALSE
)
}
\arguments{
\item{api_key}{PurpleAir API READ Key. If \code{api_key = NULL}, it
will be obtained using \code{getAPIKey("PurpleAir-read")}.}

\item{pas}{Previously generated \emph{pas} object containing \code{sensor_index}.}

\item{sensor_index}{PurpleAir sensor identifier.}

\item{startdate}{Desired start time (ISO 8601) or \code{POSIXct}.}

\item{enddate}{Desired end time (ISO 8601) or \code{POSIXct}.}

\item{timezone}{Olson timezone used to interpret dates.}

\item{correction_FUN}{Correction function applied to \code{pat} data.}

\item{verbose}{Logical controlling the generation of warning and error messages.}
}
\value{
An AirMonitor package \emph{mts_monitor} object.
}
\description{
Download, parse and enhance hourly timeseries data from PurpleAir
and create an object of class \code{mts_monitor} for use with the AirMonitor
package.
}
\note{
This is an early implementation that does not include any data QC.
}
\section{Correction}{


Data from PurpleAir is typically corrected to more closely correlate with
EPA regulatory monitors.

By default, \code{correction_FUN} uses the EPA correction equation described
in (see \link{PurpleAir_correction}). Users may supply their own
correction function as long as it meets the following two criteria:
\enumerate{
\item{It must accept an "hourly pat" object as the first argument.}
\item{It must return an augmented "hourly pat" object with an additional
\code{"pm2.5_corrected"} column in \code{pat$data}.}
}

A trivial example to just use \code{pm2.5_atm} would be:

\preformatted{
my_correction <- function(pat) {
  pat$data$pm2.5_corrected <- pat$data$pm2.5_atm
  return(pat)
}
}
}

\examples{
\donttest{
# Fail gracefully if any resources are not available
try({

library(AirSensor2)

initializeMazamaSpatialUtils()

mon <-
  PurpleAir_createMonitor(
    api_key = PurpleAir_API_READ_KEY,
    pas = MY_PAS,
    sensor_index = "76545",
    startdate = "2023-01-01",
    enddate = "2023-01-08",
    timezone = "UTC",
    verbose = TRUE
  )

}, silent = FALSE)
}
}
\references{
\href{https://www2.purpleair.com}{PurpleAir}

\href{https://api.purpleair.com}{PurpleAir API}

\href{https://www2.purpleair.com/policies/terms-of-service}{PurpleAir Terms of service}

\href{https://www2.purpleair.com/pages/license}{PurpleAir Data license}

\href{https://www2.purpleair.com/pages/attribution}{PurpleAir Data Attribution}
}
