% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PurpleAir_API.R
\docType{data}
\name{PurpleAir_PAS_METADATA_FIELDS}
\alias{PurpleAir_PAS_METADATA_FIELDS}
\title{Comma-separated list of metadata fields used to create a \emph{pas} object}
\format{
String with comma-separated field names
}
\usage{
PurpleAir_PAS_METADATA_FIELDS
}
\description{
Character string with PurpleAir field names used in
\code{pas_downloadParseRawData()}. These fields exclude all measurements
but retain many fields thay may useful for filtering. This \emph{pas} object
can then be passed on to \emph{pat} creation functions.

A metadata-only \emph{pas} object can be useful when searching for historical
data using \code{\link{pas_filter}}.

Included fields:
\preformatted{
 [1] "name"             "model"            "hardware"
 [4] "location_type"    "private"          "latitude"
 [7] "longitude"        "altitude"         "position_rating"
[10] "firmware_version" "firmware_upgrade" "uptime"
[13] "last_seen"        "last_modified"    "date_created"
}
}
\references{
\href{https://api.purpleair.com/#api-sensors-get-sensors-data}{Get Sensors Data API}
}
\keyword{datasets}
