% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PurpleAir_API.R
\docType{data}
\name{PurpleAir_PAS_MINIMAL_FIELDS}
\alias{PurpleAir_PAS_MINIMAL_FIELDS}
\title{Comma-separated list of metadata fields used to create a \emph{pas} object}
\format{
String with comma-separated field names
}
\usage{
PurpleAir_PAS_MINIMAL_FIELDS
}
\description{
Character string with PurpleAir field names used in
\code{pas_downloadParseRawData()}. These fields exclude all measurements
but retains the minimal set of metadata required to create a \emph{pas} object.
This \emph{pas} object can then be passed on to \emph{pat} creation functions.

A metadata-only \emph{pas} object can be useful when searching for historical
data using \code{\link{pas_filterDate}}.

Included fields:
\preformatted{
[1] "name"          "location_type" "latitude"
[5] "longitude"     "last_seen"     "date_created"
}
}
\references{
\href{https://api.purpleair.com/#api-sensors-get-sensors-data}{Get Sensors Data API}
}
\keyword{datasets}
