% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PurpleAir_API.R
\docType{data}
\name{PurpleAir_PAT_QC_FIELDS}
\alias{PurpleAir_PAT_QC_FIELDS}
\title{Comma-separated list of fields needed for creating QC reports.}
\format{
String with comma-separated field names
}
\usage{
PurpleAir_PAT_QC_FIELDS
}
\description{
Character string with default PurpleAir field names used in
\code{pat_downloadParaseRawData()}. These fields are sufficient for most
QC algorithms and include most of the "information and status" and
"environmental" fields.

Included fields:
\preformatted{
 [1] "rssi"         "uptime"       "pa_latency"   "memory"
 [5] "humidity"     "temperature"  "pressure"     "pm2.5_atm"
 [9] "pm2.5_atm_a"  "pm2.5_atm_b"  "pm2.5_atm_b"  "pm2.5_cf_1"
[13] "pm2.5_cf_1_a" "pm2.5_cf_1_b"
}
}
\references{
\href{https://api.purpleair.com/#api-sensors-get-sensor-history-csv}{Get Sensor History API}
}
\keyword{datasets}
