% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AirSensor2-package.R
\name{US_AQI}
\alias{US_AQI}
\title{US EPA AQI Index levels, names, colors and action text}
\description{
Official, US EPA AQI levels, names, colors and action text are provided in a
list for easy coloring and labeling.

See \code{AirMonitor::\link[AirMonitor:US_AQI]{US_AQI}} for details.
}
